requirements.txt
fastapi
uvicorn
requests
langchain_core
langchain_groq
python-dotenv
